Added Help Button to Telemetry Interface
Added Help Button to Telemetry Interface

The button added presents a modal, which contains information the user
can reference if any clarification is needed on what the interface has
to offer. A CSS class for the modal was also added to ensure it had the
correct width when opened.

Adjusted Help Button Styling

Updated style to button so it can be inline with other objects on the
navbar and hide when the browser window is small.